+++ src/editor/SceneHierarchyPanel.h
@@ -2,9 +2,14 @@
 
 #include "EditorPanel.h"
 
+struct SelectionState;
+
 class SceneHierarchyPanel : public EditorPanel
 {
 public:
-    SceneHierarchyPanel();
+    SceneHierarchyPanel(SelectionState *selection);
     void OnImGuiRender(float dt) override;
+
+private:
+    SelectionState *m_selection;
 };