+++ src/editor/SelectionState.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+
+struct SelectionState {
+    int entity_id = -1;
+    std::string entity_name;
+};