+++ src/editor/ViewportPanel.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "EditorPanel.h"
+
+class ViewportPanel : public EditorPanel
+{
+public:
+    ViewportPanel();
+    void OnImGuiRender(float dt) override;
+};